Add `QeAppWorkChain`
Add first version of the `QeAppWorkChain`, which is a wrapper work chain
for the `PwRelaxWorkChain`, `PwBandsWorkChain`, and the `PdosWorkChain`.
By combining the three work chains into one, it makes it easier to
obtain a pre-populated builder using the `get_builder_from_protocol`
method and re-use results inside the work chain.