Add `QeAppWorkChain` (#41)
* Add `QeAppWorkChain`

Add first version of the `QeAppWorkChain`, which is a wrapper work chain
for the `PwRelaxWorkChain`, `PwBandsWorkChain`, and the `PdosWorkChain`.
By combining the three work chains into one, it makes it easier to
obtain a pre-populated builder using the `get_builder_from_protocol`
method and re-use results inside the work chain.

* Add Geometry Optimization work chain

* Add bands work chain.

* update `QeAppWorkChain` to new `PdosWorkChain`

* Employ get_builder_from_protocol method:

This enables all the functionalities of the QeAppWorkChain and
simplifies the input structure.

* Fix in/outputs for the `QeAppWorkChain`

* Add kpoints/degauss input overrides

Since the QE input generator tool also allows the user to
separately set the kpoints distance and smearing value,
we need to allow this information to be transferred to the
QE app, and provided as an input for the work chain, separate
from the protocols.

Here we add two top level inputs to the QeAppWorkChain:
`kpoints_distance_override` and `degauss_override`, whose values
are used to override all scf-type calculations in the work chain
(i.e. also the `vc-relax` calculation`). This could also be dealt
in the QE app itself, but since this work chain is entirely
dedicated to the QE app, I found it easier to simply add these
top level inputs.

* Add pseudo_family input argument

* Allow to modify pseudos and kpoints distance, set resources.

* Use input parameters to rewrite default settings.

Co-authored-by: Aliaksandr Yakutovich <yakutovicha@gmail.com>
Co-authored-by: Carl Simon Adorf <carl.simon.adorf@gmail.com>