REFACTOR: defaults and widgets of the steps
Several changes are made to clean up/refactor/modularize the widgets:

* Move defaults into `.yaml` that is read and used consistently. Use
`importlib_resources` for importing this into modules.
* Try to add consistency between the order of these default in the .yaml
file and how they are ordered in several commands of ths
`SubmitQeAppWorkChainStep`.
* Rename `configs` to `settings` (save for `code`, this was included in
PR #71).
* Fix issue where _both_ `OptionsConfig` (now `AdvancedSettings`) and
`SubmitQeAppWorkChainStep` would intialize a `PseudoFamilySelector`,
leading to issues in correctly assigning the pseudos.

Note that there is a small change in UI: The Materials Settings have
been moved outside of expert mode, since the layman usually can still
understand/decide whether their structure is magnetic/insulating or not.

This also seems to have fixed the issue we were having with the UI
breaking when disabling expert mode on the "Select Codes" tab.