IMPROVE: Workflow summary
Change the workflow summary to have two tables with the "WorkChain
setings" and the "Advanced Settings". Fix several pieces in the code to
make sure the output it obtained correctly.

Also move to using jinja files and CSS to write down these somewhat
larger HTML parts that can do with some styling.