👌 IMPROVE: Step 2 Workflow tab
Rearrange the content of the "Workflow" tab in step 2 of the QEapp:

Make a separate header for "Structure" where the selection of the
geometry optimization is done via three options:

* Structure as is (`RelaxType.NONE`)
* Atomic positions (`RelaxType.POSITIONS`)
* Full geometry (`RelaxType.POSITIONS_CELL`)

The "material settings" are also moved into the `WorkChainSettings`, and
placed right below the geometry optimization choice with some more
explanation.

The next header is still "Properties", but (for) now only the band
structure can be chosen here.

Finally, the "Protocol" header and content has remained the same as
before, but is not the final thing users can choose in this tab.