🐛 FIX: `kpoints_distance` overrides
Fix two issues related to the `kpoints_distance` override:

1. Make sure that the value shown in the "Advanced settings" corresponds
   to the one of the chosen protocol.
2. Fix how the `kpoints_distance_override` is set so it matches the
   expected input of the `QeAppWorkChain.get_builder_from_protocol()`
   method.

Also make some changes to how the builder parameters are handled.
Instead of having a `builder_parameters` traitlet on the
`SubmitQeAppWorkChainStep` class, simply define a getter and setter
method.

`get_input_parameters` returns a dictionary with all the inputs
parameters from the current settings of the GUI. This is used in the
`submit` step, where these `str`/`float` types are converted into the
required types for the `QeAppWorkChain.get_builder_from_protocol()`
method. The dictionary is also added to the extras of the work chain
node, so it can be used for the report and to reconstruct the parameters
when the work chain is loaded later.

`set_input_parameters` takes in a dictionary returned by the
`get_input_parameters` method. This is used to set the default
parameters, as well as parameters of previous work chains.

Some redundant traitlets are also removed, instead simply retrieving the
corresponding parameter from the widget directly.