Automatically setup QE on localhost if not present (#142)
The installation of QuantumESPRESSO into a local conda environment is
automatically triggered in case that the codes to run pw.x on localhost are not
yet configured. This is so that QE calculations can be run on the local host on
systems that do not yet have QE pre-installed.

The automated setup routine is only attempted once, the user is informed on
error, but the next time the app is opened no further attempt is made.

The automated installation requires a conda binary to be on the user's $PATH
and is completely skipped if that is not the case, with the assumption that the
app is run in a custom user environment.