IMPROVE: Rework resource specification
Slight overhaul of the resource specification for the work chain:

* Disable the PDOS codes if the property is not calculated.
* Add a check to make sure that all codes are run on the same computer
if the PDOS is calculated.
* Allow the user to specify both #nodes and #CPUs for remote codes.
For the localhost codes the "nodes" widget is disabled. Defaults and
maxima of the number of CPUs is set based on the number of CPUs and the
default number of MPI processes.
* Allow to specify the parallelization. A default value is set to try
and avoid parallelization issues with too many MPI processes.
* Add `export OMP_NUM_THREADS=1` to `localhost` code setup.

Beside these changes, the builder is now properly updated to use the
specified resources and parallelization settings.