Limit MPI ranks for `projwfc.x` and `dos.x`
Currently the `PdosWorkChain` runs the `projwfc.x` and `dos.x`
calculations with the full number of MPI ranks specified by the user,
without any kind of parallelization. This can cause failures when the
number of MPI ranks exceeds the number of FFT slabs in the
decomposition.

Here we use the same parallelization settings for the `projwfc.x`
calculation. For the `dos.x`, there is no parallellization implemented,
but this post-processing step is typically very fast, so we simply limit
the number of MPI ranks to the value that we used to set the default
number of `npools`.