🐛 FIX: use primitive structure for PDOS `nscf` restart (#196)
When both the band structure and PDOS are calculated, the
`QeAppWorkChain` will restart the `nscf` for the `PdosWorkchain` from
the `scf` calculation in the `PwBandsWorkChain`. However, the
`PwBandsWorkChain` will also primitivize the structure via Seek-path.
This meant that the input structure for the `nscf` calculation did not
match the one that Quantum ESPRESSO finds in the `data-file-schema.xml`
for the restart, causing the calculation to crash.

Here we set the `current_structure` in the work chain context to the
primitive one obtained from SeeK-path in the `inspect_bands` step of the
outline.