FIX: Running PDOS for spin-polarised system
Currently it is not possible to calculate the PDOS for a spin-polarised
system (i.e. setting Magnetism to Ferromagnetic), since the
`QeAppWorkChain` expects the `projections` output from the
`PdosWorkChain`. However, for spin-polarised systems the
`ProjwfcCalculation` returns two outputs: `projectsion_up` and
`projections_down`.

Here we adapt the `QeAppWorkChain` to check for the `projections_up`
output, and in case it is present it will add both `projections_up` and
`projections_down` to the outputs. Otherwise it will only add the
`projections` output (both only in case the `workchain_pdos` was
actually run).

To deal with plotting the corresponding (P)DOS, we simply add the total
DOS up and down states together, and separately add the PDOS orbitals
for the up and down states to the list passed to the `widget-bandsplot`.
Since the widget currently adds all the PDOS of the same orbital types
(s, p, d, f) together, there will currently not be any differentiation
between the up and down states in the visualization. This is something
that will need to be improved on the widget side.