qeapp.yml install as package_data (#256)
qeapp.yml is loaded using `importlib` but not installed as file data which causes the issue when importing modules of qe app. 
The exception was not raised since the `qe.ipynb` is run from the library folder and the file is there. But not for the case is calling from the other than the folder.