New version for src workchain (#300)
Fix the issue that `QeAppWorkChain` is not updated with the QeApp.

We are now maintaining `setup.py` of the `QeAppWorkChain` in the `src/__init__.py` separated from the QeApp PyPI package. Every time the QeApp is released, it is supposed to bump the QE workchain version so it can be updated with the installation. The bug is the version of the workchain is hardcoded to `v1.0` and updating the qeApp will not update the workchain in the end. 
It is fixed by having the workchain version synchronize with the qeApp version. It uses the `bumpver` tool to accomplish this.