Migrate to support AiiDA 2.x (#271)
It is now migrating to AiiDA 2.x with also its new features supported such as `InstalledCode`.
The CI downstream test is added. This will fix issue #205. 
To conform with the thread-local/traitlets issue, we change also all traitlets related to threads in the code base. 
We are now getting rid of all the widgets using AiiDA node as traitlets except `NodeViewWidget` and `MinimalStructureViewer`, which will be further tested.