CI fixes and improvements (#331)
Put back uploading screenshots as artifacts. Screenshots were previously generated in the `screenshots/` directory inside the Git repo.  However, we now change ownership of the Git repo that is mounted into the testing container to user jovyan. 

Therefore, the Git repo directory is no longer writable from outside the container, which means pytest cannot write to it and screenshot files are not saved. It is workaround by giving the package folder the write access of all user inside the container.

Other fixes:

1. Do not test with multiple Python versions, tests themselves run inside the container so this was meaningles.
2. Fix the `setup.cfg`. See comment below. I needed to remove the `setupcfg-fmt` pre-commit hook, but since we plan to move to `pyproject.toml` anyway I think this is fine. 
3. Version bumps of various testing dependencies.
4. Updated `pre-commit` plugins by running `pre-commit autoupdate` manually. This PR supersecedes #323