♻️ decouple override settings of builder (#382)
We were drowned in the complex logic of `QeAppWorkChain` and its interaction with the widgets of setting workflow logics and parameters overriding those set by advanced users.
This makes adding new parameters to QeApp error-prone and involves multiple files across the repo.

Meanwhile, it lacks a systematic approach to test and check the correctness of the setting workchain inputs from widgets. The developers have to actually trigger the workchain from GUI and check the QE inputs in the remote computers to find out the issues in workcahin builder settings.

The new design of how the frontend widget communicates with backend worflows is described in https://github.com/aiidalab/aiidalab-qe/issues/383. 

With the new design, the workchain triggering code is separated into three key components, the view which are widgets for parameters and inputs settings and output represented, the controller which creates the builder for workchain process and parses the report information from workchain and display in the output viewer widgets.

The setting widgets are decoupled from each other. Complex are all tucked into the builder creator method.
In this refactoring, the complexity of logic is only in terms of how the builder is created from different settings. There is another layer of complexity that is the logic of workchain itself.