Hardcode the library version when installing pseudo family (#425)
There is an issue that when the aiida-pseudo update to have the latest version of library, such as from `v1,2` to `v1.3`, since in the QeApp we hardcode the library for workchain protocol, the install command in `sssp.py` will fail because it trys to install the latest version which is `v1.3` but it not compatible with the validate version that is `v1.2`.

In this commit, we also hardcode the version to the install command to make sure when a new version is available in aiida-pseudo, it will still install the correct version.