Pseudo setter override builder properly as pw override parameters (#445)
To override pseudos in the `PwBaseWorkChain`, the pseudos dict need to be set for the `pw` namespace of the WorkChain. 
It was put as the input of the workchain and thus not take effect to override. 
In the commit, the pseudos dict is set for the `pw` namespace of the base workchain.