Pass label to WorkChainSelector (#473)
The workchain selector is very slow since it loads workchain node to retrieve information from it. This is not necessary since we can store the workchain-related information in the submission step and pass it to workchain selector. It accelerates the loading time of workchain selector in QeApp and makes the loading of the app much faster.