Move docker stack build from docker stack to aiidalab-qe (#449)
I adapted most of the codes I used in `aiidalab-docker-stack`. Except in this repo, I call docker buildx bake directly instead of using the `doit` script.
It reveals that with this implementation, the integration test can be easily run with the image build. Also, provide a preview container on `ghcr` for immediate testing of new feature implementation.
The CI test for ARM64 is run on the self-hosted runner deployed in PSI. 

Meanwhile, I moved the notebook test to use the built image directly.