Fix/remove workchain output (#510)
Fix #390

The electronic structure panel is still hardcoded because it needs the results from both bands and pdos plugins. When we find a good solution to support results from multiple plugin, we can remove both the code in the electronic structure result and the code in the workchain.

For the moment, we remove the code in the workchain, but still hardcoded the electronic structure result panel.