Adding dynamic visualization of codes using block layout (#544)
This commit enables the possibility to dynamically show codes in step 3, even if something is changed in step 2 after the confirmation. If you confirm again with a different set of properties, the shown code is updated, at variance with before.
Moreover, codes that are not shown will not occupy blank space in the app, instead, they are now in the main branch. I achieved this using display instead of visibility.

---------

Co-authored-by: Miki Bonacci