Install pseudos from downloaded files (#524)
fixes #490

In this commit, we pinging the aiida-pseudo version to use the
--download-only option to download the pseudo archive to the container
for quick installation when the container first time start.
Note that from the test all 8 pseudos seems to only save ~10s but in the place where the network condition is bad, this is very helpful.

Further more Refactoring the methods in setup_pseudos.py to support flexible parameter passing.