XAS: Refactor CH Pseudo System
Refactors the system for importing core-hole pseudopotentials from
using groups to using a `yaml` file to define which pseudos should
be present in the ch pseudos archive. Greatly simplifies the logic
for checking for the presence of pseudopotentials and populating
lists of available pseudos.

Yaml file `pseudo_toc` is obtained from the Core_Level_Spectra_Pseudos
archive (https://github.com/PNOGillespie/Core_Level_Spectra_Pseudos)
and must be updated on both the archive and the plugin when new
pseudos are available on the CLS archive.

Also adds a step in `workchain.py` to check for mismatches in
which pseudos should be present and re-download the archive file
if there is a mismatch. Added in order to check for inconsistencies
and to automatically update the archive if `pseudo_toc.yaml` is
updated with new pseudos.