confirm steps when loading previous result (#570)
In https://github.com/aiidalab/aiidalab-qe/pull/537 we introduce a mechanism that when the structure setup step or configure setup step is modified without confirmation, the submission step will be blocked. When loading finished calculations from the workflow selection bar, since setup steps are not confirmed, the user has to go back to previous steps to confirm which is not ideal. 
This PR make it automatically confirm the setup steps when the workflows are loaded from workflow selection dropdown.