XAS Plugin: Various Improvements
Adds a range of improvements and fixes to the plugin.
Fixes:
* Fixes an issue where the `Setting` panel of the plugin would
cause a crash due to a variable being re-defined during runtime.
* Adds a `.close()` statement to the end of the HTTP request when
accessing the pseudo archive file for download.
Additions:
* The app now retrieves all contributions to the final spectrum for the
selected element and plots each together. The spectra are aligned in terms of
energy in the same way as the `XspectraCrystalWorkChain` uses to generate the
final spectrum. Each contributing spectrum is scaled according to the ratio of
site multiplicity to total multiplicity. The provided broadening tools will apply
to all spectra on the widget simultaneously.
* Adds a button to the `Result` panel to download a CSV file of the spectra
presented on the Plotly widget. This will download a file titled
`{element}_XAS_Spectra.csv` which contains all spectra on the widget (total
spectrum and all site contributions). For sites with a multiplicity ratio
different from 1, the weighted and unweighted spectra are printed in separate
columns.