New bandspdoswidget for band structure and pdos rendering
The old bandsplot widget https://github.com/aiidalab/widget-bandsplot was separately
maintained in the repo with complex and not well modulized JS code.
In this PR, the new bandspdoswidget is introduced to have it all implemented in python
and in ipywidget framework with using plotly as plot engine.
The output band structure is in a publish ready level with a flexible control on the
groups of orbitals/atoms user want to projected.