Add Feature: XAS Plugin (#580)
Adds a plugin to calculate XANES spectra using the XspectraCrystalWorkChain of AiiDA-QE. The Setting panel allows users to simply select which element to calculate XAS for, as well as tuning the supercell size to be used in the calculation. This panel also provides the option to change the type of core-hole approximation between FCH (removing the excited electron with tot_charge = 1) and XCH (allowing the excited electron to occupy the conduction band. The result will display both the complete powder XAS for the input structure and the contributions from each symmetrically non-equivalent site. 


---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Xing Wang <xingwang1991@gmail.com>