Add Feature: Documentation Page for XAS Plugin (#614)
This PR adds a complete documentation set for using the XAS plugin of AiiDALab-QE, covering both calculation setup and post-processing of the final spectra using the Plotly-derived widget provided with the plugin. The instructions should adequately cover all features of the XAS App using a self-contained, robust, and well-tested example.