Add plugin: X-ray Photoelectron Spectroscopy (XPS) (#518)
This PR adds a plugin for calculating X-ray Photoelectron Spectroscopy (XPS) spectra using the XpsWorkChain of aiida-quantumespresso package. The Setting panel allows users to select the core-level (element and orbital) to be calculated. The availability of core-evel in the panel depends on the corresponding pseudopotentials.

The Result panel displays the final XPS spectrum
- Broadened spectrum using a Voight profile (combined Lorenzian and Gaussian).
- The chemical shift values (differences in total energy relative to the lowest value)
- Absolute binding energy
- The Offset Energies (δ)

To make the setting as simple as possible for user. The following default core_hole_treatment are used:
- `full` for molecule
- `xch_smear` for metal
- `xch_fixed` for insulator

The supercell_min_parameter is updated based on protocol.

---------


Co-authored-by: Peter Gillespie <55498719+PNOGillespie@users.noreply.github.com>