add a `update_inputs` to plugin's `workchain_and_builder` (#656)
This PR enhances the plugin API by introducing a new feature to the `workchain_and_builder` configuration: the `update_inputs` method. 

This addition addresses limitations in the previous design, where the main workchain rigidly passed only the `structure` to the plugins' workchains. This approach had notable disadvantages:

- It mandated that all plugins' workchains accept an input named `structure`, leading to errors if this was not adhered to. This is the case in the newly developed plugin: [aiidalab-qe-hp](https://github.com/superstar54/aiidalab-qe-hp).
- It restricted data passed to plugins to be hardcoded, thereby limiting plugin developers' ability to tailor the inputs based on their specific requirements.

The introduction of the `update_inputs` method provides a dynamic mechanism for updating the inputs of plugins' workchains. It accepts two parameters: `inputs` and `ctx`. This method enables developers to customize how the inputs to a plugin's workchain are updated, leveraging the broader context of the Workchain. An immediate benefit of this improvement is observed in the **bands** and **pdos** plugins, which can now access `current_number_of_bands` from the context if available.