Add bumpver pre-commit hook
This bumpver is configured to automatically run
bumpver_pre_commit.sh script before it makes the release commit.
The script checks for common mistakes, such as making a release commit
in a wrong branch, or trying to push to a wrong remote.

This is an initial implementation that should be
made more robust in the future.