Feature/post install plugins (#755)
* Adding the post installation trigger when we install a plugin.
Putting also the post install information for the vibroscopy in the plugin.yaml file.

* Modification of the name of the post_install command for vibroscopy GUI.
Documentation will come in the next commit in order to help developers
to set the post installation for their plugins.

Also, updating of the plugin_list notebook, now the post installation step
title is also printed.

to test with the aiidalab-qe-vibroscopy plugin, you should modify the data dictionary
of the first block of the notebook like this:

data['aiidalab-qe-vibroscopy']['github'] = "https://github.com/mikibonacci/aiidalab-qe-vibroscopy@automatic_installation_codes"
data['aiidalab-qe-vibroscopy']['post_install'] = "setup-phonopy"

* Modifying the plugin_list notebook to take the local plugins.yaml file.
I think this is better, as the file can also contain specific plugin versions
which are compatible with the current local qeApp version (not with the remote main).

Also, due to this, I modified the plugin.yaml file to update the branch of the vibroscopy
plugin. (will change to the release version soon).

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Adding the documentation section on how to setup the post-install for a given plugin.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Updating dependencies: allowing installation of aiidalab-widgets-base[optimade]<2.4.0

* Updating the plugin_list notebook following the pre-commit suggestion.

* documentation on automatic post install plugins refined and ready.

* Changing the aiidalab-qe-vibroscopy installation from github to pip

* Removing the attempt of fetching plugins.yaml file from remote.
We should keep it local so we don't risk to load incompatible plugins (wrt the app version).

* Adding a space in post_install command in the plugin_list notebook.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>