Add button to trigger the update of the results tabs. (#868)
The process monitor updates the result tabs in another thread. This leads to an issue: when the user interacts with the result, the widget runs in the thread differently from the one running in the process monitor. Therefore, triggering the automatic updating of the result tabs is not possible.

In this PR, we don't use process monitor anymore. Instead, we add a button to let the user trigger the update manually. I also modified the layout of the buttons (kill, clean) to make the UI pretty.