Implement MVC design and lazy loading (#802)
This is a large-scale redesign that focuses on the separation of data/business logic from the UI. Data is contained in `Model` classes as `traitlets` traits, which are observed in the `Controller`/`View` widgets and handled by `Controller` event handlers. `Controller`s also observe changes in their respective `View`s to handle user events. The design relies predominantly on the Observer pattern, with a few exceptions handled as a Mediator pattern.

Lazy loading is implemented via `render` methods triggered when the user visits the relevant component. However, the `Model` network is decoupled from the UI and remains in sync as a source of truth. `View`s sync up with the `Model` network on render.

Some changes in the UI were included here to better fit with the MVC design and the principles of lazy loading. Clear documentation will be provided in a follow-up PR.