Implement path for in-app guides (#763)
This PR adds a mechanism for adding in-app guides meant as example use-cases for users. The mechanism leverages a global instance of the `GuideManager`, which tracks the current selected guide in an `active_guide` traitlet, which is observed by all `InAppGuide` instances. Instances are placed throughout the app with unique identifiers or with hard-coded widgets and associated `guide_id`. Guide placeholders are spread throughout the app. Additional placeholders may be added by plugin developers within their defined panels.