Install aiida-pseudos from repo to avoid download issue (#1028)
Since `aiida-pseudo` need `aiida-core` > 2.6, however, we still want to support the previous aiida-core version, so in the image, we use the lastest aiida-pseudo to download the pseudos and install it. After the pseudos are installed, it can be used by the App. This is a temporary solution to install aiida-pseudos from repo to avoid downloading issues.