Define structure results panel based on relaxation property (#1015)
This PR implements a structure results panel that is always shown. The tab title and source node are dependent on the nature of the calculation as follows:
- No relaxation - title = "Initial structure"; source = `node.inputs.structure`
- Some relaxation - title = "Relaxed structure"; source = `node.outputs.structure`

Auto rendering is adjusted to ensure the initial structure panel is not dependent on having results (there are none) but instead is always auto rendered.