Replace advanced settings override checkbox with a reset button (#1012)
This PR discards the override mechanism in the advanced settings panel. All widgets are enabled and may be overridden by the user. A reset button is added in place of the override checkbox to reset the advanced settings to their defaults w.r.t the basic settings and selected structure.