Redesign summary view (#1044)
This PR introduces a bullet list alternative format for the summary report (no actual bullets). The format is defined in `qeapp.yml` as `summary_format: list` and can be overridden by the user in a config file (see #1007 for info regarding file). `summary_format` takes list or table.

Styles are adjusted such that the data download widget is to the right of the summary instead of below. This is true for either selected format.

Lastly, the PR adds "Workflow properties" and "Structure properties" sections to the summary.