Unify electronic structure result panels (#1039)
This PR implements a single unified panel for electronic structure results. The title of the panel's tab is dynamic depending on the selected properties. In addition, if both bands and PDOS are selected, selection controls are added at the top of the panel for the user to choose which property they would like to plot.