Use plugin utility to set default starting magnetization (#1041)
This PR uses the QE plugin `get_starting_magenetization` utility function to set default initial magnetic moments for the structure when magnetism is turned on. The defaults are set to the default `magmom` provided by the utility scaled by the Z valence value for the element defined in the selected pseudopotential. Note that for those elements of a zero `magmom` default, a general default magnetization of 0.1 is returned.

If multiple tags are provided for the same species, a warning is displayed providing some instructions for antiferromagnetic configurations.