Implement logging widget (mostly for devs) (#1133)
This PR adds a flag, which if True, adds an `Output` widget at the bottom of the app, accompanied by a reset button. The widget can then be passed down to any part of the app that needs to be tested.
The PR hardcodes one use-case - monitoring. The `ProcessMonitor` was recently changed (https://github.com/aiidalab/aiidalab-widgets-base/pull/669) to accept an output widget and use it to log exceptions in monitored callbacks. This PR passes the output widget (if included) down to the `ProcessMonitor` instance, exposing its exceptions. This is useful for debugging issues in the results step.