Tetrahedra occupation for molecules  (#1140)
fix: Use `tetrahedra_lin` instead of `tetrahedra_opt` for molecular projections  

`tetrahedra_opt` does not produce projections in `projwfc` for molecular systems. This update replaces it with `tetrahedra_lin`, ensuring proper projection calculations while maintaining compatibility with periodic systems.  

Additionally, a fix is change to the test to search for the bands property by looking for a child with "Electronic Band Structure" info.