Hubbard: change int values of eigenvalues to float  (#1178)
The values for starting_ns_eigenvalue should be floats, not integers.

Previously, the app had two issues:

It was sending integers instead of floats.
It assigned spin-up as 0 and spin-down as 1, which is not the convention in QE.
Additionally, the test was incorrectly checking spin-down as 1.

This PR updates the values to floats and corrects the tests accordingly.