Fix maximum value in selection of atomic indices  (#1251)
The atomic indices selection in the edit atom tags panel
supports the generation of slices, e.g. `1..5`. However,
the maximum needs to be smaller less than or equal to the number of
sites in the system. With this fix, we also allow larger
indices but simply ignore them. This reproduces the behavior
of lists in Python.