Fix default magnetic moments for species with tags (#1252)
Currently, the default magnetic moments do not properly work when
the atoms have tags assigned. This is for example the case when modelling
antiferromagnetic systems.

The problem was that we first populate `self._default_magnetic_moments`
using the utility methods of the AiiDA-QE plugin. Those expect the chemical
elements without any tags. Hence, when dealing with tags, `self._default_magnetic_moments`
simply contained the elements with tags and 0.1 as magnetization. In the
`_to_moment` we were correctly using the `kind.symbol`. However, we first
tried to retrieve the values from `self._default_magnetic_moments`, which
doesn't cotain them as outlined before, again resulting in the default
magnetization of 0.1.