The “latest” tag is created by the manifest job (#1290)
In both the GHCR and DockerHub metadata steps (for both build jobs), the `latest` tag line has been removed. Now, only the release tag (e.g. `v25.04.0`) is generated. The `latest` tag is then created by the manifest job (if the tag is a release and does not contain `alpha`).