Use shallow copy to avoid creating a new AiiDA node every time
The parameters collected from the GUI are passed to the builder of the QeAppWorkChain, and then passed to each plugin's WorkChain builder. In order to avoid each plugin mute orginal parameters, we used deepcopy before. But deepcopy will create new AiiDA nodes, which break the provenance. In this PR, we replace the deepcopy by shallow copy, so the plugin can not mute the original dict.