Use shallow copy to avoid creating a new AiiDA node every time (#1341)
The parameters collected from the GUI are passed to the builder of the QeAppWorkChain, and then passed to each plugin's WorkChain builder. To avoid each plugin muting original parameters, we used deepcopy before. However, deepcopy will create new AiiDA nodes, breaking the provenance. In this PR, we replace the deepcopy with a shallow copy, so the plugin can not mutate the original dict.