Fix PP upload widget (#1348)
This PR addresses several issues with the pseudopotential upload widget. 
It implements a system of warnings w.r.t invalid uploads.
If a valid, non-built-in pseudo is uploaded, the `functional`, `library`, and `family` are set to `None`.
This required support in the plugin now provided in `aiida-quantumespresso v4.12.0`.
Magnetization-to-moments conversion now uses `UpfData` nodes to extract Z values.
Lastly, the summary is updated to show selected pseudopotentials and reflect custom states.