Inject default codes into `ResourceSettingsModel` (#1368)
Default codes defined in `qeapp.yml` were not actually considered. The app 
was selecting the first fetched code per entry point. This PR correct this behavior.
In the process, it also corrects callback registration for code model traits prior to 
rendering, which was also not present. On the tests front, it corrects test code
handling, as well as switches to using the pytest fixtures from AiiDA Core  2.6.