Avoid storing nodes in frontend - use UUIDs (#1389)
We already use the process UUID in the results step to avoid DB session conflicts across threads. This PR extends this principle throughout the app, allowing a broader use of threads in the future. In particular, the `HasInputStructure` and `HasProcess` mixins now operate with uuids. A property is introduced to each referencing the node.